refactor(workout-editor): extract sortable row + add-button helper
WorkoutPlanEditor.tsx was juggling its own state, dnd-kit shell, the
sortable row implementation, the editor header, the empty state, and
two near-identical "+ Add exercise" CTAs. Pulled the biggest of these
out:

- SortableRow → components/workout-logging/WorkoutPlanEditorRow.tsx
  (owns dnd-kit wiring + the layered current-row cues)
- AddExerciseButton local helper folds the footer and empty-state
  CTAs into a single component with a `block` size variant

Backdrop dismissal now sits on a real `<button>` so the a11y rule that
fired during pre-commit (useKeyWithClickEvents) is satisfied without
suppressions; semantically the backdrop is a labelled close target.
